requirements.txt
psycopg2-binary
PyYAML
pandas
sqlalchemy